requirements.txt
python-telegram-bot==21.10
groq==0.18.0
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.12
gTTS==2.5.4
openai==1.63.2
flask==3.0.3
